check relation before modify

+++ src/main/java/com/abc/crm/service/CompanyService.java
@@ -2,25 +2,33 @@ package com.abc.crm.service;
 
 import com.abc.crm.controller.dto.req.CompanyReqDto;
 import com.abc.crm.controller.dto.res.CompanyResDto;
+import com.abc.crm.dao.ClientDao;
 import com.abc.crm.dao.CompanyDao;
 import com.abc.crm.service.bo.CompanyBo;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
 
 import java.sql.Timestamp;
 import java.time.Instant;
 import java.util.Optional;
 
+@Log4j2
 @Service
 public class CompanyService {
 
     private final CompanyDao companyDao;
+    private final ClientDao clientDao;
 
-    public CompanyService(CompanyDao companyDao) {
+    public CompanyService(
+            CompanyDao companyDao,
+            ClientDao clientDao) {
         this.companyDao = companyDao;
+        this.clientDao = clientDao;
     }
 
     public Optional<CompanyResDto> getOne(Long id) {
-        throw new UnsupportedOperationException();
+        return companyDao.getOne(id)
+                .map(CompanyResDto::valueOf);
     }
 
     public Optional<Long> addOne(CompanyReqDto companyReqDto, String username) {
@@ -47,6 +55,10 @@ public class CompanyService {
     }
 
     public boolean deleteOne(Long id) {
+        if (clientDao.getByCompanyId(id).size() > 0) {
+            log.debug("id={} is in use.", id);
+            return false;
+        }
         return companyDao.deleteOne(id);
     }
